Update unit and integration tests with flattened module
