+++ scripts/test-cli-integration
@@ -11,15 +11,16 @@ key_b=$(prio keygen)
 shared=$(prio shared-seed)
 SHARED_SECRET=$shared
 
+SHARED_SECRET=${SHARED_SECRET:-$(jq -r ".shared_seed" <<<$shared)}
 FILENAME=${FILENAME:-"data.ndjson"}
 N_DATA=${N_DATA:-3}
 BATCH_ID=${BATCH_ID:-"test"}
 
-SERVER_A_PUBLIC_KEY=${SERVER_A_PUBLIC_KEY:-$(jq -r ".public_key" <<< $key_a)}
-SERVER_B_PUBLIC_KEY=${SERVER_B_PUBLIC_KEY:-$(jq -r ".public_key" <<< $key_b)}
+SERVER_A_PUBLIC_KEY=${SERVER_A_PUBLIC_KEY:-$(jq -r ".public_key" <<<$key_a)}
+SERVER_B_PUBLIC_KEY=${SERVER_B_PUBLIC_KEY:-$(jq -r ".public_key" <<<$key_b)}
 
-SERVER_A_PRIVATE_KEY=${SERVER_A_PRIVATE_KEY:-$(jq -r ".private_key" <<< $key_a)}
-SERVER_B_PRIVATE_KEY=${SERVER_B_PRIVATE_KEY:-$(jq -r ".private_key" <<< $key_b)}
+SERVER_A_PRIVATE_KEY=${SERVER_A_PRIVATE_KEY:-$(jq -r ".private_key" <<<$key_a)}
+SERVER_B_PRIVATE_KEY=${SERVER_B_PRIVATE_KEY:-$(jq -r ".private_key" <<<$key_b)}
 
 CLIENT_BUCKET=${CLIENT_BUCKET:-"working/client"}
 SERVER_A_BUCKET=${SERVER_A_BUCKET:-"working/server_a"}
@@ -31,15 +32,17 @@ SERVER_B_BUCKET=${SERVER_B_BUCKET:-"working/server_b"}
 
 if [[ "$FILENAME" == "data.ndjson" ]]; then
 
-    cat << EOF > ${CLIENT_BUCKET}/data.ndjson
-[1, 0, 0]
-[1, 1, 0]
-[1, 1, 1]
+    jq -c '{payload: .}' <<EOF >${CLIENT_BUCKET}/${FILENAME}
+[1, 0, 0, 0, 0]
+[1, 1, 0, 0, 0]
+[1, 1, 1, 0, 0]
+[1, 1, 1, 1, 0]
+[1, 1, 1, 1, 1]
 EOF
 
     jq -c '.' ${CLIENT_BUCKET}/${FILENAME}
 
-    python3 -m prio encode-shares \
+    prio encode-shares \
         --n-data ${N_DATA} \
         --batch-id ${BATCH_ID} \
         --public-key-hex-internal ${SERVER_A_PUBLIC_KEY} \
@@ -56,12 +59,12 @@ fi
 # verify1
 ###########################################################
 
-python3 -m prio verify1 \
+prio verify1 \
     --n-data ${N_DATA} \
     --batch-id ${BATCH_ID} \
     --server-id A \
     --private-key-hex ${SERVER_A_PRIVATE_KEY} \
-    --shared-secret $shared \
+    --shared-secret $SHARED_SECRET \
     --public-key-hex-internal ${SERVER_A_PUBLIC_KEY} \
     --public-key-hex-external ${SERVER_B_PUBLIC_KEY} \
     --input ${SERVER_A_BUCKET}/raw/${FILENAME} \
@@ -73,7 +76,7 @@ cp \
     ${SERVER_A_BUCKET}/intermediate/internal/verify1/${FILENAME} \
     ${SERVER_B_BUCKET}/intermediate/external/verify1/
 
-python3 -m prio verify1 \
+prio verify1 \
     --n-data ${N_DATA} \
     --batch-id ${BATCH_ID} \
     --server-id B \
@@ -94,18 +97,18 @@ cp \
 # verify2
 ###########################################################
 
-python3 -m prio verify2 \
+prio verify2 \
     --n-data ${N_DATA} \
     --batch-id ${BATCH_ID} \
     --server-id A \
     --private-key-hex ${SERVER_A_PRIVATE_KEY} \
-    --shared-secret $shared \
+    --shared-secret $SHARED_SECRET \
     --public-key-hex-internal ${SERVER_A_PUBLIC_KEY} \
     --public-key-hex-external ${SERVER_B_PUBLIC_KEY} \
     --input ${SERVER_A_BUCKET}/raw/${FILENAME} \
     --input-internal ${SERVER_A_BUCKET}/intermediate/internal/verify1/${FILENAME} \
     --input-external ${SERVER_A_BUCKET}/intermediate/external/verify1/${FILENAME} \
-    --output ${SERVER_A_BUCKET}/intermediate/internal/verify2/ \
+    --output ${SERVER_A_BUCKET}/intermediate/internal/verify2/
 
 jq -c '.' ${SERVER_A_BUCKET}/intermediate/internal/verify2/${FILENAME}
 
@@ -113,18 +116,18 @@ cp \
     ${SERVER_A_BUCKET}/intermediate/internal/verify2/${FILENAME} \
     ${SERVER_B_BUCKET}/intermediate/external/verify2/
 
-python3 -m prio verify2 \
+prio verify2 \
     --n-data ${N_DATA} \
     --batch-id ${BATCH_ID} \
     --server-id B \
     --private-key-hex ${SERVER_B_PRIVATE_KEY} \
-    --shared-secret $shared \
+    --shared-secret $SHARED_SECRET \
     --public-key-hex-internal ${SERVER_B_PUBLIC_KEY} \
     --public-key-hex-external ${SERVER_A_PUBLIC_KEY} \
     --input ${SERVER_B_BUCKET}/raw/${FILENAME} \
     --input-internal ${SERVER_B_BUCKET}/intermediate/internal/verify1/${FILENAME} \
     --input-external ${SERVER_B_BUCKET}/intermediate/external/verify1/${FILENAME} \
-    --output ${SERVER_B_BUCKET}/intermediate/internal/verify2/ \
+    --output ${SERVER_B_BUCKET}/intermediate/internal/verify2/
 
 jq -c '.' ${SERVER_B_BUCKET}/intermediate/internal/verify2/${FILENAME}
 
@@ -136,12 +139,12 @@ cp \
 # aggregate
 ###########################################################
 
-python3 -m prio aggregate \
+prio aggregate \
     --n-data ${N_DATA} \
     --batch-id ${BATCH_ID} \
     --server-id A \
     --private-key-hex ${SERVER_A_PRIVATE_KEY} \
-    --shared-secret $shared \
+    --shared-secret $SHARED_SECRET \
     --public-key-hex-internal ${SERVER_A_PUBLIC_KEY} \
     --public-key-hex-external ${SERVER_B_PUBLIC_KEY} \
     --input ${SERVER_A_BUCKET}/raw/${FILENAME} \
@@ -155,18 +158,18 @@ cp \
     ${SERVER_A_BUCKET}/intermediate/internal/aggregate/${FILENAME} \
     ${SERVER_B_BUCKET}/intermediate/external/aggregate/
 
-python3 -m prio aggregate \
+prio aggregate \
     --n-data ${N_DATA} \
     --batch-id ${BATCH_ID} \
     --server-id B \
     --private-key-hex ${SERVER_B_PRIVATE_KEY} \
-    --shared-secret $shared \
+    --shared-secret $SHARED_SECRET \
     --public-key-hex-internal ${SERVER_B_PUBLIC_KEY} \
     --public-key-hex-external ${SERVER_A_PUBLIC_KEY} \
     --input ${SERVER_B_BUCKET}/raw/${FILENAME} \
     --input-internal ${SERVER_B_BUCKET}/intermediate/internal/verify2/${FILENAME} \
     --input-external ${SERVER_B_BUCKET}/intermediate/external/verify2/${FILENAME} \
-    --output ${SERVER_B_BUCKET}/intermediate/internal/aggregate/ \
+    --output ${SERVER_B_BUCKET}/intermediate/internal/aggregate/
 
 jq -c '.' ${SERVER_B_BUCKET}/intermediate/internal/verify2/${FILENAME}
 
@@ -178,12 +181,12 @@ cp \
 # publish
 ###########################################################
 
-python3 -m prio publish \
+prio publish \
     --n-data ${N_DATA} \
     --batch-id ${BATCH_ID} \
     --server-id A \
     --private-key-hex ${SERVER_A_PRIVATE_KEY} \
-    --shared-secret $shared \
+    --shared-secret $SHARED_SECRET \
     --public-key-hex-internal ${SERVER_A_PUBLIC_KEY} \
     --public-key-hex-external ${SERVER_B_PUBLIC_KEY} \
     --input-internal ${SERVER_A_BUCKET}/intermediate/internal/aggregate/${FILENAME} \
@@ -192,12 +195,12 @@ python3 -m prio publish \
 
 jq -c '.' ${SERVER_A_BUCKET}/processed/${FILENAME}
 
-python3 -m prio publish \
+prio publish \
     --n-data ${N_DATA} \
     --batch-id ${BATCH_ID} \
     --server-id B \
     --private-key-hex ${SERVER_B_PRIVATE_KEY} \
-    --shared-secret $shared \
+    --shared-secret $SHARED_SECRET \
     --public-key-hex-internal ${SERVER_B_PUBLIC_KEY} \
     --public-key-hex-external ${SERVER_A_PUBLIC_KEY} \
     --input-internal ${SERVER_B_BUCKET}/intermediate/internal/aggregate/${FILENAME} \
