+++ tests/resources/cli/client/data.ndjson
@@ -1,0 +1,5 @@
+{"payload":[1,0,0,0,0]}
+{"payload":[1,1,0,0,0]}
+{"payload":[1,1,1,0,0]}
+{"payload":[1,1,1,1,0]}
+{"payload":[1,1,1,1,1]}
