+++ tests/resources/cli/config.json
@@ -1,0 +1,4 @@
+{
+    "n_data": 5,
+    "batch_id": "test"
+}
