+++ tests/resources/cli/server_a/intermediate/external/aggregate/data.ndjson
@@ -1,0 +1,1 @@
+"AZWrRLClnH6BVNEhGPWrXlo8bE6FuRNoiiKrKF/0u6hYxk3pytSrKYEQ2Wp+wiZug0SrZFmiWnlc8I+nCAs="
