+++ tests/resources/cli/server_a/processed/data.ndjson
@@ -1,0 +1,1 @@
+{"id": "4dcd62a3-58be-4d75-8815-42aea4578fbc", "timestamp": "2020-07-03T00:36:38.440536", "payload": [5, 4, 3, 2, 1]}
