+++ tests/resources/cli/server_a/raw/data.ndjson
@@ -1,0 +1,5 @@
+{"id": "efcbc73e-10ab-451f-9677-f497e077f16f", "payload": "t9EmP6cNaRV4KqEfea+6F6354qLI81Rso+Vb0UpgvV1wHRIplYVE1tmPmHWO3QyQSiCPnO7pHL+s7W44sUEiZ1yfpn5a6egxFtk3RjdOSXGJa6a2uxKo7A6hXaUhZejh5SeyAjAJkTnN7gUBilYezAIoRq5Dl3RjPU42FY9fA1UgSUMjkN/3YlVTacM6TinsbJ6VncSUKh8MohJ+/YywxaoCWC7/19RuLCUlzL8Ki5dnqCQ7HwjOIO4Y43TqDlb9iUyAmEiAHLI630ZPJDz4VcAdGv2QvtXNW5BmTGfayTVT/s4pMzeCJLWnw1QV1Kl81LiEUN43tzQO8DrujYXUvuHGf98bTh4FvzLL8dBoBlSZOslC8XWpUXv2wulAb0ZVgMoVEK8wbc8="}
+{"id": "7db6f738-c8c5-49e8-a136-2b66a21aa2ad", "payload": "0Ta5FiZhwreH6+YGs5RjLsmmU+w3v9uDLS1sym9CiFy7n1hKZCno04FIFw36yugYGjzAvt8HLfVPSHppWYixdVIXHfmRxk7METuwSwHuNiXgTR8sxAspjzWVei/gViLrLHJ3RahOjP5pTtIjb8qT/5CraU24CtwhNdLWLhus7hgopMtLdoecQLh8yoJ2y8C4r2wlFYtsYxz7c1nETILqgfB54zR0H+bZGw2XK34NpAfbrk1t+vYCDUMcvCDwerr/+5EJahV6SFYHvTTccHbWrQQ2ZajRy78vwZk7Wbf5c6MMOsBDWWmBeCbh84hh+Bwf8aKMyFkqwqoTkv4u9CjsRtcpBj2ILJMqLNzORkw5xN3ZcK10tJhhaiRZP4xx+F5Iv5tATDJaFBs="}
+{"id": "820ffbe6-b255-4426-8196-00f4b5454332", "payload": "TAmv1GB2yicKgr87iWiqXCkvtyjYEruJQ2BkdIhNjB9NtuioC8sdekVwYeF6F+UJoC+yTECRkqVsAwF453gYWYlq+cB8aI5b8Yj1PsbcPTL+sRe3VRsQHAabk9S7SdXizL6t8gPETpM+cGqjh5DwWVwYzBpMpSOxU45y6maFuKXXB4xIPYwR0sdO8Y5TS7s4xbTjiOMKpyhHjqexA+ZvmY6S7pnMsk6EyiWTy7+Bdqy20YFxnerxN2sXo5xoj5Puiuxa9VaSfa4edKrdNN9gDK5nyES1sV2K8N2t9VUlTqOLz1ZPSH3nQs9f1N1YCrjV67woM/cs7B2+PR6BO+6WivMAyPlC3SV7XvEaJs9BlSi14JMoKuGxwnlt7Bt4/qM2hNDW7uxJhg=="}
+{"id": "b90ad1d9-3e57-4316-a34d-62b7caf1c042", "payload": "59RiqrMKtNw/kKu66OHvwlIinYA9Bs6kGwCRaKxmZUsoRqlIW+8KEJ2WU81PTtad4ouNJeABzpVyslshCRlutYLloWCockNv7ANKUNdc0YiMw33YaujgxlUVXJYu52UyI4B8nNfLmMB17O5ps391sTj2MT429SoyWsSljoJmjSwiDqpN8fNBrdzIzzlZ/WpyoS8LB9QO7IhNcZgmumXFmiHEbZnQh6vyFxUJ0sSEaqn+TJnviUbNYa0K1M3enlWRqCJsoukG520lFuWBIm7zh0f2H2CuFTlIJ6xYNSvwDL1SpXMpbwEBMXInbhneoJAnCUq35NlBYm37x/rkL3rx/35yA+8dG+7j92AoT74qSqfpukKaucXhvnIxAaLbig1bbJrVFylzf64="}
+{"id": "0b19ced1-23bd-4ff8-afd8-88970a9cd994", "payload": "XQOJ6wrOYICk8Idvm4ZclVHgtjWuNNloem4d3TOWwjzsJcvwYui2DGB7gun23E3KbtrEwZ9Nh4wti2GvkE3IfpIe7uPTYeuG4oFycre5/6DSd2BJQXjvkRqf9DEW9c+k9+sopEpMb1kKovV0fhJmFXA68Ld429vLRcXvNgTXzgzHgmmNC/O5ZaKXUDp5vh+lz0jBF4nK97aKbbecq/jJSyrXr5rR9EMWPQUa0UmRVG9SuWkiVPLkvijg+A10Y7B+jDMMQ64YTro/on7b0Adj4JYjf9SxmTOhMQIvjS4orExxZiNQqvWLJcS0ZnGnctggBzwogu6DsbMMoZr5esjWyYvn7YSyWnTk4ZVqvIWJCgDYVaDbgz7MZzFESQh+mSLMuJvK44UBXoQ="}
