+++ tests/resources/cli/server_a_keys.json
@@ -1,0 +1,4 @@
+{
+  "private_key": "E1A78115B45F30CC300125945757BB41F4341AA604B2C68909DA92EA059D55DE",
+  "public_key": "7FD03F1BAFC080675DFA6E0BDCDD06044F6517AF35D6B5F4F8A6717056F9B20B"
+}
