+++ tests/resources/cli/server_b/intermediate/external/aggregate/data.ndjson
@@ -1,0 +1,1 @@
+"AJWrO09aY4F+qy7m5xGrIaXDk7F6RuyfdeOrV6ALRFenObIeNTCrVn7vJpWBPdmZfL+rG6ZdpYajD3Bg9/c="
