+++ tests/resources/cli/server_b/processed/data.ndjson
@@ -1,0 +1,1 @@
+{"id": "e5562a36-ca47-4c0b-9b75-38f4a2265f91", "timestamp": "2020-07-03T00:36:38.713909", "payload": [5, 4, 3, 2, 1]}
