+++ tests/resources/cli/server_b/raw/data.ndjson
@@ -1,0 +1,5 @@
+{"id": "efcbc73e-10ab-451f-9677-f497e077f16f", "payload": "qDqKTEj+ERK71e4q5eXeTWms3YMSQSgLhALeU+5Ftx+4YGw6cGZ8R7bqozxb3k5LX2wV1tK69dxGTk2bB+VzOefQ7YanbzPVN6Nm+sulOPG4RLKwcl0AE1eUxOAQIlpz1eeZswshm4o2Ung571rbCSDPfLnDwA/Uz3R7P3HbR2wK1hyFVyGGF2gT3U02aI6nUWVfDUJDe7/kZ5A="}
+{"id": "7db6f738-c8c5-49e8-a136-2b66a21aa2ad", "payload": "kuunwmOLzNSoMmEcmofZWMVVGomD2/0XKQE5ppJ+MQTmXHyn6DVZMDHVJ7c2TVClgyD38Mc+m0/zjxdEkuW+U8nbMA/TD1k2gejSfSu85OjajR9wKe1SpO5C8tC544RZPkvaRRS6Ew3LdrfMGduKTMW+DKSWfmCM0rIoa+SmWXMKgRGRYEXfMp8baIZPzmmwDKTR3vIPaFFhWMA="}
+{"id": "820ffbe6-b255-4426-8196-00f4b5454332", "payload": "ZttTTGJVYbao4KvC06JAgb3SYh5/qfa/bmaq2KbTo3vZ2DD6zUgku9vGycos4sWbgDtZyr+iBxBmzfEl6um7tVVsE5LJJpARROA+G0whZ+D/PFsHzmIuGO9R5ElXwJ7SkVCjnmpNbLFDIKA571x8RVCJJyfR/PZihl0c/AOqPtYnOMxB8lxqWNJZQNhjMZDZP4cHPonQJXkz2JI="}
+{"id": "b90ad1d9-3e57-4316-a34d-62b7caf1c042", "payload": "RgrPChZQnc3djQh/4KVbNLwhNTbGjIdpbG/lk6r/rFukldQ+4kpvJ4h7iDlXSrPdnp2DVFu0JYELMzUlADd4EsmQQxvmjqG3MdJAYa93c26ajtYtzUo+7+OHjHuGXUJrh+OhqxvfXVHtBO3PWI1C9eX6nima2bllrExFTMUEzE1NT29Pf7E4BboVR3grv5c9h+L/IIyRSHdeUV4="}
+{"id": "0b19ced1-23bd-4ff8-afd8-88970a9cd994", "payload": "3LPPP3wghpCwv1kBAP+5veiTDQiXebwDnK+nwq5WQQhg0i8HW3w/AQLQP44fBa2KXaJB8WVTwfgPmTR1LKI0qWdQYkQJsZ/mFbQoYaLaRcNgjRqtAMdp8B/0GRxFzurBiGHHYR6tjExefdDbUDcJRCOsIUK9Z7nJbVVqqoqOPK6AvegMEpFNqv0MU++ejSfmNl7aQP70AnHasLA="}
