+++ tests/resources/cli/server_b_keys.json
@@ -1,0 +1,4 @@
+{
+  "private_key": "6D5D37297BCE4ADA2B885FB325C06379E24CF16E487BD8E3501F972661625ED2",
+  "public_key": "56EA4A77FC70699374594999E558ED9FD27B394EFB45C0A862D2B8276B40407B"
+}
