+++ tests/resources/cli/shared_seed.json
@@ -1,0 +1,3 @@
+{
+  "shared_seed": "elYuGM7pVBq1a3LG4Y5KfQ=="
+}
